Upgrade to opentracing-java 0.21.0 (#137)
 and handle PEER_PORT change from short to int

+++ jaeger-core/src/test/java/com/uber/jaeger/SpanTest.java
@@ -261,7 +261,7 @@ public class SpanTest {
   @Test
   public void testSpanDetectsEndpointTags() {
     int expectedIp = (127 << 24) | 1;
-    short expectedPort = 8080;
+    int expectedPort = 8080;
     String expectedServiceName = "some-peer-service";
     Span span = (Span) tracer.buildSpan("test-service-operation").start();
     Tags.PEER_HOST_IPV4.set(span, expectedIp);
